Refactored website generation under WebsiteArtifactory
Reduced the complexity of the giant `generate_website` function.
Separated individual tabs of site into their own calls for scope
encapsulation.
Aim is to make it easier to chase down subtle issues, such as why we
don't have pagination arrows for the retrospective tab.

Also, separating configuration from the generation call with attributes
on the object reduces the number of args that must be passed through to
called methods, and avoids the workaround used previously of functions
defined within the function to take advantage of its scope.

Addresses #141, #149.